Renamed elements to layout
