+++ client/src/pages/Profile.js
@@ -2,9 +2,9 @@ import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import ProfileDisplay from "body/ProfileDisplay";
 
-import Header from "elements/Header";
-import Page from "elements/Page";
-import Body from "elements/Body";
+import Header from "layout/Header";
+import Page from "layout/Page";
+import Body from "layout/Body";
 
 const useStyles = makeStyles((theme) => ({
   //
